コンポーネント分け
